+++ gsoap/custom/struct_timeval.h
@@ -1,7 +1,7 @@
 /*
 	struct_timeval.h
 
-	Custom serializer for struct timeval as xsd::dateTime
+	Custom serializer for struct timeval as xsd:dateTime
 
 	Because time_t (binds to xsd:dateTime) lacks fractional seconds, struct
 	timeval can be used to represent microseconds since 1970-01-01.