+++ gsoap/src/soapcpp2.h
@@ -44,8 +44,8 @@ A commercial use license is available from Genivia, Inc., contact@genivia.com
 #include "error2.h"
 
 #ifndef VERSION
-# define VERSION "2.8.82" /* Current version */
-# define GSOAP_VERSION 20882
+# define VERSION "2.8.83" /* Current version */
+# define GSOAP_VERSION 20883
 #endif
 
 #ifdef WIN32