+++ gsoap/stdsoap2.cpp
@@ -1,5 +1,5 @@
 /*
-        stdsoap2.c[pp] 2.8.82
+        stdsoap2.c[pp] 2.8.83
 
         gSOAP runtime engine
 
@@ -52,7 +52,7 @@ A commercial use license is available from Genivia, Inc., contact@genivia.com
 --------------------------------------------------------------------------------
 */
 
-#define GSOAP_LIB_VERSION 20882
+#define GSOAP_LIB_VERSION 20883
 
 #ifdef AS400
 # pragma convert(819)   /* EBCDIC to ASCII */
@@ -86,10 +86,10 @@ A commercial use license is available from Genivia, Inc., contact@genivia.com
 #endif
 
 #ifdef __cplusplus
-SOAP_SOURCE_STAMP("@(#) stdsoap2.cpp ver 2.8.82 2019-03-14 00:00:00 GMT")
+SOAP_SOURCE_STAMP("@(#) stdsoap2.cpp ver 2.8.83 2019-04-18 00:00:00 GMT")
 extern "C" {
 #else
-SOAP_SOURCE_STAMP("@(#) stdsoap2.c ver 2.8.82 2019-03-14 00:00:00 GMT")
+SOAP_SOURCE_STAMP("@(#) stdsoap2.c ver 2.8.83 2019-04-18 00:00:00 GMT")
 #endif
 
 /* 8bit character representing unknown character entity or multibyte data */
@@ -250,9 +250,11 @@ static size_t frecv(struct soap*, char*, size_t);
 static int tcp_init(struct soap*);
 static const char *tcp_error(struct soap*);
 
+#if !defined(WITH_IPV6)
+static int tcp_gethost(struct soap*, const char *addr, struct in_addr *inaddr);
+#endif
 #if !defined(WITH_IPV6) || defined(WITH_COOKIES)
 static int tcp_gethostbyname(struct soap*, const char *addr, struct hostent *hostent, struct in_addr *inaddr);
-static int tcp_gethost(struct soap*, const char *addr, struct in_addr *inaddr);
 #endif
 
 static SOAP_SOCKET tcp_connect(struct soap*, const char *endpoint, const char *host, int port);
@@ -562,7 +564,9 @@ static int tcp_done = 0;
 #endif
 
 #if (defined(_AIX43) || defined(TRU64) || defined(HP_UX)) && defined(HAVE_GETHOSTBYNAME_R)
+#ifndef h_errno
 extern int h_errno;
+#endif
 #endif
 
 /******************************************************************************/
@@ -1868,31 +1872,19 @@ int
 SOAP_FMAC2
 soap_binary_search_string(const char **a, int n, const char *s)
 {
-  int i, k;
-  for (i = 1, k = n; k > 1; k >>= 1)
-    i <<= 1;
-  k = i >> 1;
-  i--;
-  for (;;)
+  int min = 0, max = n-1;
+  while (min <= max)
   {
-    if (i >= n)
-    {
-      i -= k;
-    }
+    int mid = (min+max)/2;
+    int r = strcmp(s, a[mid]);
+    if (r < 0)
+      max = mid - 1;
+    else if (r > 0)
+      min = mid + 1;
     else
-    {
-      int r = strcmp(a[i], s);
-      if (r == 0)
-        return i;
-      if (r > 0)
-        i += k;
-      else
-        i -= k;
-    }
-    if (k == 0)
-      return -1;
-    k >>= 1;
+      return mid;
   }
+  return -1;
 }
 
 /******************************************************************************/
@@ -5074,7 +5066,7 @@ tcp_error(struct soap *soap)
 static int
 tcp_gethostbyname(struct soap *soap, const char *addr, struct hostent *hostent, struct in_addr *inaddr)
 {
-#if defined(__GLIBC__) && (!_GNU_SOURCE && !defined(_POSIX_C_SOURCE) && !defined(_XOPEN_SOURCE) && defined(HAVE_GETHOSTBYNAME_R)) || _POSIX_C_SOURCE >= 200112L || _XOPEN_SOURCE >= 600 || defined(__ANDROID__) || (defined(HAVE_GETHOSTBYNAME_R) && (defined(__GNU__) || defined(__GNUC__))) || (defined(HAVE_GETHOSTBYNAME_R) && (defined(FREEBSD) || defined(__FreeBSD__)))
+#if defined(__GLIBC__) && (!_GNU_SOURCE && !defined(_POSIX_C_SOURCE) && !defined(_XOPEN_SOURCE) && defined(HAVE_GETHOSTBYNAME_R)) || _POSIX_C_SOURCE >= 200112L || _XOPEN_SOURCE >= 600 || defined(__ANDROID__) || (defined(HAVE_GETHOSTBYNAME_R) && (defined(FREEBSD) || defined(__FreeBSD__)))
   int r;
   char *tmpbuf = soap->tmpbuf;
   size_t tmplen = sizeof(soap->tmpbuf);
@@ -5102,7 +5094,7 @@ tcp_gethostbyname(struct soap *soap, const char *addr, struct hostent *hostent, 
       return SOAP_OK;
     }
   }
-#if defined(__GLIBC__) && (!_GNU_SOURCE && !defined(_POSIX_C_SOURCE) && !defined(_XOPEN_SOURCE) && defined(HAVE_GETHOSTBYNAME_R)) || _POSIX_C_SOURCE >= 200112L || _XOPEN_SOURCE >= 600 || defined(__ANDROID__) || (defined(HAVE_GETHOSTBYNAME_R) && (defined(__GNU__) || defined(__GNUC__))) || (defined(HAVE_GETHOSTBYNAME_R) && (defined(FREEBSD) || defined(__FreeBSD__)))
+#if defined(__GLIBC__) && (!_GNU_SOURCE && !defined(_POSIX_C_SOURCE) && !defined(_XOPEN_SOURCE) && defined(HAVE_GETHOSTBYNAME_R)) || _POSIX_C_SOURCE >= 200112L || _XOPEN_SOURCE >= 600 || defined(__ANDROID__) || (defined(HAVE_GETHOSTBYNAME_R) && (defined(FREEBSD) || defined(__FreeBSD__)))
   while ((r = gethostbyname_r(addr, hostent, tmpbuf, tmplen, &hostent, &soap->errnum)) < 0)
   {
     if (tmpbuf != soap->tmpbuf)
@@ -5594,15 +5586,19 @@ again:
   {
     if (inet_pton(AF_INET6, soap->client_interface, res->ai_addr) != 1)
     {
-      soap->errnum = soap_socket_errno(sk);
-      freeaddrinfo(ressave);
-      soap_set_receiver_error(soap, tcp_error(soap), "inet_pton() failed in tcp_connect()", SOAP_TCP_ERROR);
-      soap->fclosesocket(soap, sk);
-      soap->client_interface = NULL;
-      return soap->socket = SOAP_INVALID_SOCKET;
+      if (inet_pton(AF_INET, soap->client_interface, res->ai_addr) != 1)
+      {
+        soap->errnum = soap_socket_errno(sk);
+        freeaddrinfo(ressave);
+        soap_set_receiver_error(soap, tcp_error(soap), "inet_pton() failed in tcp_connect()", SOAP_TCP_ERROR);
+        soap->fclosesocket(soap, sk);
+        soap->client_interface = NULL;
+        return soap->socket = SOAP_INVALID_SOCKET;
+      }
     }
     soap->client_interface = NULL; /* disable client interface, so need to set it again before the next connect */
   }
+#ifndef WITH_LEAN
   if ((soap->omode & SOAP_IO_UDP))
   {
     if (soap_memcpy((void*)&soap->peer.storage, sizeof(soap->peer.storage), (const void*)res->ai_addr, res->ai_addrlen))
@@ -5615,6 +5611,7 @@ again:
     freeaddrinfo(ressave);
     return sk;
   }
+#endif
 #endif
 #ifndef WITH_LEAN
   if (soap->connect_timeout)